+++ CHANGELOG.md
@@ -5,7 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No changes yet.
+### Added
+- Added `Name` option for `Provide` to add named values to the container
+  without rewriting constructors. See package documentation for more
+  information.
+
+### Changed
+- `name:"..."` tags on nested Result Objects are no longer ignored.
 
 ## [1.3.0] - 2017-12-04
 ### Changed